Enclosed selenium-standalone dependency for test environment.

+++ src/utils/gulp/gulp-tasks-test.js
@@ -66,36 +66,42 @@ module.exports = function(gulp, options) {
     }
   });
 
-  var selenium = require('selenium-standalone');
+  var selenium;
 
   gulp.task('integration:clean', function () {
-    selenium.child.kill();
+    if (selenium) {
+      selenium.child.kill();
+    }
   });
 
   gulp.task('selenium', function(done) {
-
-    selenium.install({
-      logger: function() {}
-    }, function(err) {
-      if (err) {
-        return done(err);
-      }
-
-      selenium.start(function(err, child) {
+    if (options.e2ePaths) {
+      selenium = require('selenium-standalone');
+      selenium.install({
+        logger: function() {}
+      }, function(err) {
         if (err) {
           return done(err);
         }
 
-        if (process.env.TRAVIS) {
-          child.stderr.on('data', function(data) {
-            console.log(data.toString());
-          });
-        }
+        selenium.start(function(err, child) {
+          if (err) {
+            return done(err);
+          }
 
-        //saving the child to kill it later (oops)
-        selenium.child = child;
-        done();
+          if (process.env.TRAVIS) {
+            child.stderr.on('data', function(data) {
+              console.log(data.toString());
+            });
+          }
+
+          //saving the child to kill it later (oops)
+          selenium.child = child;
+          done();
+        });
       });
-    });
+    } else {
+      console.log('You need options.e2ePaths to start the selenium server.');
+    }
   });
 };
